+++ packages/core/src/components/x-data-provider-storage/storage/service.ts
@@ -30,6 +30,8 @@ export class StorageService implements IDataProvider, IDataMutator {
   }
 
   async set(key: string, value: string) {
+    const existing = await this.get(key)
+    if (existing == value) return
     this.localStorage?.setItem(this.prefix + key, value)
     this.eventBus.emit(DATA_EVENTS.DataChanged, {
       provider: this.name,
